implementar las funcionalidades de cargar imagen y mostrar dichas imagenes para el administrador

+++ src/main/java/com/curso/ecommerce/service/UploadFileService.java
@@ -12,12 +12,20 @@ import org.springframework.web.multipart.MultipartFile;
 @Service
 public class UploadFileService {
 
-    private String folder = "images//";
+    private String folder = "images" +File.separator;
 
     public String saveImage(MultipartFile file) throws IOException{
         if(!file.isEmpty()){
+            //Crear carpeta sino existe
+            File directorio = new File(folder);
+
+            if(!directorio.exists()){
+                directorio.mkdirs();
+            }
+
             byte[] bytes = file.getBytes();
             Path path = Paths.get(folder+file.getOriginalFilename());
+            System.out.println("Ruta: " +path.toString());
             Files.write(path, bytes);
             return file.getOriginalFilename();
         }
@@ -25,9 +33,11 @@ public class UploadFileService {
     }
 
     public void deleteImage(String nombre){
-        String ruta = "image//";
-        File file = new File(ruta+nombre);
-        file.delete();
+        File file = new File(folder + nombre);
+
+        if(file.exists()){
+            file.delete();  
+        }
     }
 
 }
